Create (dysfunctional) main menu
My fancy pants tool scripts that update pickups and enemies in realtime with a tool script setter uhh...

Present the original crash when the scene changes from a different one. That is, if I load the debug level scene and play it directly, everything works fine. But if I change to that scene from the main menu, we're back to square one on all my `Callable#call_deferred()` business. How lovely.